Apply suggestions from code review - move `maybePrintSimpleScalacOutput` back to `ScalaCommand` - remove ScalacLikeCommand trait - remove the `-X` alias for cross-compilation - remove unnecessary exit status code verification in Run command ITs
