+++ modules/cli-options/src/main/scala/scala/cli/commands/CompileCrossOptions.scala
@@ -5,7 +5,6 @@ import caseapp._
 // format: off
 final case class CompileCrossOptions(
   @HelpMessage("Cross-compile sources")
-  @ExtraName("X")
   @Hidden
     cross: Option[Boolean] = None
 )
