+++ modules/cli-options/src/main/scala/scala/cli/commands/ScalacOptions.scala
@@ -17,6 +17,7 @@ final case class ScalacOptions(
     scalacOption: List[String] = Nil
 )
 // format: on
+
 object ScalacOptions {
 
   private val scalacOptionsArg = Arg("scalacOption")
