+++ website/docs/reference/cli-options.md
@@ -213,8 +213,6 @@ Available in commands:
 
 #### `--cross`
 
-Aliases: `-X`
-
 Cross-compile sources
 
 ## Coursier options
@@ -263,8 +261,6 @@ Available in commands:
 <!-- Automatically generated, DO NOT EDIT MANUALLY -->
 
 #### `--cross`
-
-Aliases: `-X`
 
 ## Default file options
